Per-SNR calibration, float32 dataset, and SNR-vs-metric plots
detector/evaluate_structured_dae.py: replace the single global threshold
with one 95th-percentile threshold per SNR level, computed from that SNR's
own clean validation scores and applied only to that SNR's test rows.
Results table now also reports raw counts (n_flagged/n_total, n_fp/n_clean)
alongside the rates.

detector/generate_frozen_dataset.py: dataset_X was being cast to float64
(411MB); switched to float32 (~215MB) since nothing downstream needs the
extra precision.

Regenerated results/dataset_*.npy (gitignored, not committed) and reran the
evaluation: per-SNR calibration measurably separates non_adaptive from
clean at higher SNR (e.g. SNR 30: fpr=0.033 vs detection_rate=0.967), while
detection_rate still equals fpr for the adaptive attacker at every SNR
(ROC-AUC ~0.50-0.52 throughout) — the ceiling-clipped adaptive attacker
remains statistically indistinguishable from clean traffic to this
detector, calibration notwithstanding.

Added results/roc_auc_vs_snr.png and results/detection_rate_vs_snr.png
(150 DPI) plotting both attacker types against SNR.

+++ detector/generate_frozen_dataset.py
@@ -45,7 +45,7 @@ def build_dataset() -> tuple[np.ndarray, np.ndarray, np.ndarray]:
 
             scenario_seed += 1
 
-    X = np.stack(X_rows).astype(np.float64)
+    X = np.stack(X_rows).astype(np.float32)
     y = np.asarray(y_rows, dtype=np.int64)
     snr = np.asarray(snr_rows, dtype=np.int64)
     return X, y, snr
